borrar chofer y combi pero bien

+++ app/Http/Controllers/ChoferController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Chofer;
+use App\Models\Combi;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreChoferes;
 
@@ -32,6 +33,9 @@ class ChoferController extends Controller
     }
 
     public function eliminarChofer(Chofer $chofer){
+        $combi = Combi::where('chofer_id', '=', $chofer->id)->get()->first();
+        if (!empty($combi))
+          return redirect()->route('combi19.listarChoferes');
         $chofer->delete();
         return redirect()->route('combi19.listarChoferes');
     }
